Task/ampoffline 1017/resources in af

+++ app/modules/repository/RepositoryManager.js
@@ -5,6 +5,8 @@ import Logger from '../util/LoggerManager';
 
 const logger = new Logger('RepositoryManager');
 
+const DO_NOT_DELETE = ['.gitignore'];
+
 /**
  * Repository Manager
  *
@@ -36,15 +38,59 @@ const RepositoryManager = {
   },
 
   /**
+   * Deletes files and directories that are not found under the specified tree of used content paths
+   * @param usedPathsTree use tree of content paths
+   */
+  cleanupUnusedContent(usedPathsTree) {
+    this._didCleanUp(usedPathsTree);
+  },
+
+  _didCleanUp(usedTree, ...pathParts) {
+    const currentPath = (pathParts && pathParts.length && pathParts[pathParts.length - 1]) || null;
+    let canDelete = currentPath && (!usedTree || !usedTree.has(currentPath)) && !DO_NOT_DELETE.includes(currentPath);
+    try {
+      const fullPath = FileManager.getFullPath(REPOSITORY_DIR, ...pathParts);
+      const stat = FileManager.statSyncFullPath(fullPath);
+      if (!stat) {
+        logger.warn(`Path not found: ${fullPath}`);
+      } if (stat.isFile()) {
+        // FILE
+        if (canDelete) {
+          FileManager.deleteFile(fullPath);
+        }
+      } else {
+        // DIR
+        const filesOrDirs = FileManager.readdirSync(REPOSITORY_DIR, ...pathParts);
+        let isEmpty = !filesOrDirs.length;
+        if (filesOrDirs.length) {
+          const subTree = currentPath ? ((usedTree && usedTree.get(currentPath)) || null) : usedTree;
+          isEmpty = !filesOrDirs.filter(fileOrDir => !this._didCleanUp(subTree, ...pathParts, fileOrDir)).length;
+        }
+        if (canDelete && isEmpty) {
+          FileManager.rmdirSync(REPOSITORY_DIR, ...pathParts);
+        }
+      }
+    } catch (error) {
+      logger.error(error);
+      canDelete = false;
+    }
+    return canDelete;
+  },
+
+  /**
    * Stores a local file to the repository directory
    * @param srcFilePath
+   * @param sync if to copy the file synchronously or provide a promise
    * @return the content metadata
    */
-  storeLocalFileToRepository(srcFilePath) {
+  storeLocalFileToRepository(srcFilePath, sync) {
     this.init(false);
     const content = this._buildContent(srcFilePath);
-    FileManager.copyDataFileSync(srcFilePath, REPOSITORY_DIR, content[PATH]);
-    return content;
+    if (sync) {
+      FileManager.copyDataFileSync(srcFilePath, REPOSITORY_DIR, content[PATH]);
+      return content;
+    }
+    return FileManager.copyDataFileAsync(srcFilePath, REPOSITORY_DIR, content[PATH]).then(() => content);
   },
 
   _buildContent(srcFilePath) {
@@ -75,7 +121,7 @@ const RepositoryManager = {
    * @return {*|string}
    */
   getFullContentFilePath(content) {
-    return FileManager.getFullPath(REPOSITORY_DIR, this._getContentPath(content));
+    return content && FileManager.getFullPath(REPOSITORY_DIR, this._getContentPath(content));
   },
 
   /**
@@ -88,12 +134,41 @@ const RepositoryManager = {
   },
 
   /**
+   * Attempts to delete safely a content from repository files
+   * @param content
+   */
+  attemptToDeleteContent(content) {
+    logger.info('attemptToDeleteContent');
+    try {
+      RepositoryManager.deleteFromRepository(content);
+    } catch (error) {
+      const tmpPath = RepositoryManager.getFullContentFilePath(content);
+      logger.error(`Could not properly cleanup temporary content or its folders (${tmpPath}): "${error}". 
+          A new atempt will be done later by the cleanup task.`);
+    }
+  },
+
+  /**
    * Deletes a file from the repository
    * @param content the file metadata
    */
   deleteFromRepository(content) {
-    const fullFilePath = FileManager.getFullPath(REPOSITORY_DIR, this._getContentPath(content));
+    const relativePath = this._getContentPath(content);
+    if (!relativePath) {
+      return;
+    }
+    const fullFilePath = FileManager.getFullPath(REPOSITORY_DIR, relativePath);
     FileManager.deleteFile(fullFilePath);
+
+    let dirsToCheck = FileManager.splitPath(relativePath).filter(p => p);
+    while (dirsToCheck && dirsToCheck.length > 1) {
+      dirsToCheck.pop();
+      if (FileManager.readdirSync(REPOSITORY_DIR, ...dirsToCheck).length) {
+        dirsToCheck = null;
+      } else {
+        FileManager.rmdirSync(REPOSITORY_DIR, ...dirsToCheck);
+      }
+    }
   },
 
   /**
@@ -103,7 +178,7 @@ const RepositoryManager = {
    * @private
    */
   _getContentPath(content: Object) {
-    return content[PATH];
+    return content && content[PATH];
   },
 
 };
